+++ pkg/workload/workloadimpl/random.go
@@ -23,7 +23,7 @@ func RandStringFast(rng rand.Source, buf []byte, alphabet string) {
 	// that point, you likely should consider PrecomputedRand.
 	alphabetLen := uint64(len(alphabet))
 	// floor(log(math.MaxUint64)/log(alphabetLen))
-	lettersCharsPerRand := uint64(math.Log(float64(math.MaxUint64)) / math.Log(float64(alphabetLen)))
+	lettersCharsPerRand := uint64(int64(math.Log(float64(math.MaxUint64)) / math.Log(float64(alphabetLen))))
 
 	var r, charsLeft uint64
 	for i := 0; i < len(buf); i++ {
